refactor: reduce tagBloom to 4 bytes, ensure 4-byte alignment for headers
- Reduce tagBloom from uint64_t to uint32_t in TrackHotHeader (36 bytes
total)
- Fix TrackColdHeader padding for 4-byte alignment (48 bytes total)
- Add splitInt64/combineInt64 helpers in ByteView.h for int64 splitting
- Update TrackView accessors to use combineInt64 for fileSize/mtime
- Add 4-byte padding in serializeHot() and encodeColdData()
- Add size alignment assertions in TrackStore write methods
- Update tests for new struct sizes

+++ include/rs/expr/ExecutionPlan.h
@@ -114,7 +114,7 @@ namespace rs::expr
     std::vector<std::string> stringConstants;
 
     // Bloom filter for tag fast-path rejection
-    std::uint64_t tagBloomMask = 0;
+    std::uint32_t tagBloomMask = 0;
 
     // If true, the query matches all tracks (no conditions)
     bool matchesAll = false;
